Fix GCC error
There was an error from comparing unsigned and signed. Also, localize
the string creation to the error paths and force the use of braces to
align with style on if-else statements.

+++ glslang/Include/arrays.h
@@ -201,7 +201,7 @@ struct TSmallArrayVector {
 
     const TArraySize& operator[](int index) const
     {
-        assert(sizes && index < sizes->size());
+        assert(sizes && index < (int)sizes->size());
         return (*sizes)[index];
     }
